+++ BusinessLayer/Interfaces/ISortingService.cs
@@ -2,5 +2,5 @@ namespace BusinessLayer.SortingAlgorithms;
 
 public interface ISortingService
 {
-    List<int> Sort(List<int> intNumbers);
+    List<double> Sort(List<double> intNumbers);
 }
